sslh-main: fix config_lookup_int call
This func takes an int, not a long.  The current code triggers a warning:

sslh-main.c: In function 'config_parse':
sslh-main.c:275:5: warning: passing argument 3 of 'config_lookup_int' from incompatible pointer type [enabled by default]
     if (config_lookup_int(&config, "timeout", &timeout) == CONFIG_TRUE) {
                                               ^
In file included from sslh-main.c:26:0:
/usr/include/libconfig.h:266:12: note: expected 'int *' but argument is of type 'long int *'
 extern LIBCONFIG_API int config_lookup_int(const config_t *config,

+++ sslh-main.c
@@ -254,7 +254,7 @@ static int config_protocols(config_t *config, struct proto **prots)
 static int config_parse(char *filename, struct addrinfo **listen, struct proto **prots)
 {
     config_t config;
-    long int timeout;
+    int timeout;
     const char* str;
 
     config_init(&config);
